requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.7.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
loguru==0.7.3
lxml==5.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.1
Pygments==2.18.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.3.1
pypng==0.20220715.0
pyppeteer==2.0.0
pytest==8.2.2
python-bidi==0.6.6
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.3.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
svglib==1.5.1
tinycss2==1.4.0
tomli==2.2.1
tqdm==4.67.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ujson==5.10.0
uritools==4.0.3
urllib3==1.26.20
uvicorn==0.30.1
watchfiles==0.22.0
webencodings==0.5.1
websockets==10.4
win32_setctime==1.2.0
xhtml2pdf==0.2.17
zipp==3.20.2
pyodbc==4.0.30
pymssql==2.1.5